PR feedback - decouple `copyToClipboard` from the component
Decouple `copyToClipboard` from the component and make it a utility function that executes a callback instead
